refactor: split admin.rs (1,619 lines) into admin/ subdirectory
Decompose into 4 focused modules by domain responsibility:
- admin/backup.rs (248 lines) — BackupManager impl
- admin/maintenance.rs (385 lines) — MaintenanceManager impl  
- admin/welcome.rs (377 lines) — WelcomeProvider impl
- admin/stats.rs (200 lines) — StatsProvider impl

Zero API changes — all pub symbols remain accessible via same paths.
