+++ src/memory_core/storage/sqlite/admin/mod.rs
@@ -1,0 +1,16 @@
+//! Admin sub-modules for the SQLite storage backend.
+//!
+//! This module aggregates four logical areas of administrative functionality:
+//!
+//! - [`backup`] — database snapshot creation, rotation, listing, and restore
+//! - [`maintenance`] — health checks, consolidation, compaction, and session clearing
+//! - [`welcome`] — session-start context assembly (`WelcomeProvider`)
+//! - [`stats`] — aggregated analytics (`StatsProvider`)
+//!
+//! All impl blocks target [`super::SqliteStorage`] and are kept in separate
+//! files purely for readability — the public API is unchanged.
+
+mod backup;
+mod maintenance;
+mod stats;
+mod welcome;
